Feat: useGetFollowCount, useGetCheckFollowing
